Module Omd_parser

Beware: the functions in this module may raise exceptions! If you use them, you should be careful.

type r = Omd_representation.t

accumulator (beware, reversed tokens)

and p = Omd_representation.tok list

context information: previous elements

and l = Omd_representation.tok list

tokens to parse

and main_loop = ?⁠html:bool -> r -> p -> l -> Omd_representation.t

most important loop, which has to be given as an argument

val default_parse : ?⁠extensions:Omd_representation.extensions -> ?⁠default_lang:string -> l -> Omd_representation.t

Translate tokens to Markdown representation.

parameter lang

language for blocks of code where it was not specified. Default: "".

module type Env = sig ... end

module Default_env : functor (Unit : sig ... end) -> Env

module Make : functor (Env : Env) -> sig ... end