Parameter Make.1-Env

val rc : Omd_representation.ref_container

reference container

val extensions : Omd_representation.extensions

reference container

list of parser extensions

val default_lang : string

list of parser extensions

default language for code blocks

val gh_uemph_or_bold_style : bool

default language for code blocks

flag: bold/emph using using underscores is by default github-style, which means that underscores inside words are left as underscore, rather than special characters, because it's more convenient. However it is also less expressive because then you can't bold/emph a part of a word. You might want to set this flag to false.

val blind_html : bool

flag: bold/emph using using underscores is by default github-style, which means that underscores inside words are left as underscore, rather than special characters, because it's more convenient. However it is also less expressive because then you can't bold/emph a part of a word. You might want to set this flag to false.

flag: if true, will not check whether a used HTML tag actually exists in HTML.

val strict_html : bool

flag: if true, will not check whether a used HTML tag actually exists in HTML.

flag: if true, will only accept known inline HTML tags in inline HTML.

val warning : bool

flag: if true, will only accept known inline HTML tags in inline HTML.

flag: if true, will output warnings

val warn_error : bool

flag: if true, will output warnings

flag: if true, will convert warnings to errors